+++ packages/back-end/src/routers/saved-group/saved-group.router.ts
@@ -3,19 +3,14 @@ import z from "zod";
 import { wrapController } from "../wrapController";
 import { validateRequestMiddleware } from "../utils/validateRequestMiddleware";
 import * as rawSavedGroupController from "./saved-group.controller";
+import {
+  postSavedGroupBodyValidator,
+  putSavedGroupBodyValidator,
+} from "./saved-group.validators";
 
 const router = express.Router();
 
 const savedGroupController = wrapController(rawSavedGroupController);
-
-export const postSavedGroupBodyValidator = z.object({
-  groupName: z.string(),
-  owner: z.string(),
-  type: z.enum(["condition", "list"]),
-  condition: z.string().optional(),
-  attributeKey: z.string().optional(),
-  values: z.string().array().optional(),
-});
 
 router.post(
   "/",
@@ -24,13 +19,6 @@ router.post(
   }),
   savedGroupController.postSavedGroup
 );
-
-export const putSavedGroupBodyValidator = z.object({
-  groupName: z.string().optional(),
-  owner: z.string().optional(),
-  values: z.string().array().optional(),
-  condition: z.string().optional(),
-});
 
 router.put(
   "/:id",
