+++ packages/back-end/types/saved-group.d.ts
@@ -1,8 +1,8 @@
-import z from "zod";
+import { z } from "zod";
 import {
   postSavedGroupBodyValidator,
   putSavedGroupBodyValidator,
-} from "../src/routers/saved-group/saved-group.router";
+} from "../src/routers/saved-group/saved-group.validators";
 
 /**
  * @deprecated
